refactor: renaming runtime helper to jsonmerge
We are renaming the `runtime` helper package to `jsonmerge` to convey
it's single utility.

In addition I also realized that `fix-imports` was not working since by
default it skips generated files. So to avoid issues I created a small
fix-imports helper dedicated for this file.
