+++ lib/accessgraph/apiclient/jsonmerge/jsonmerge.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package runtime
+package jsonmerge
 
 import (
 	"bytes"
